完善功能

+++ eam/clearlib.sh
@@ -8,7 +8,7 @@ echo "conf_file:$conf_file";
 
 maven_dir=`cat $conf_file|grep ${hostname}.maven_dir|awk -F "=" '{print $2}'`
 echo "maven_dir:$maven_dir"
- [[ -d "$maven_dir/com/github/foxnic" ]];then
+if [[ -d "$maven_dir/com/github/foxnic" ]];then
   cd $maven_dir/com/github/foxnic
   echo "start to clear foxnic maven lib";
   rm -rf *
